+++ src/types/LossFunction.ts
@@ -1,0 +1,5 @@
+
+export enum LossFunction {
+  MSE = 'meanSquaredError',
+  CCE = 'categoricalCrossentropy'
+}