+++ src/types/Optimizer.ts
@@ -1,0 +1,4 @@
+export enum Optimizer {
+  ADAM = 'adam',
+  SGD = 'sgd'
+}