+++ src/types/TrainingParams.ts
@@ -1,0 +1,11 @@
+import { LossFunction } from './LossFunction';
+import { Optimizer } from './Optimizer';
+
+export interface TrainingParams {
+  learningRate: number;
+  optimizer: Optimizer;
+  loss: LossFunction;
+  validationSplit: number;
+  batchSize: number;
+  epochs: number;
+}